backend/requirements.txt
# Python 3.11+
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-cov==6.1.0
httpx==0.28.1
pyjwt==2.12.1
passlib==1.7.4
bcrypt==5.0.0
fastapi-cache2==0.2.2
mypy==1.15.0
ruff==0.11.5
